move op_to_string to qldpc.math
