+++ qldpc/circuits_test.py
@@ -22,6 +22,7 @@ import pytest
 import stim
 
 from qldpc import circuits, codes, external
+from qldpc.math import op_to_string
 from qldpc.objects import Pauli
 
 
@@ -45,17 +46,17 @@ def test_state_prep() -> None:
 
         # the state of the simulator is a +1 eigenstate of code stabilizers
         for row in code.get_stabilizer_ops():
-            string = circuits.op_to_string(row)
+            string = op_to_string(row)
             assert simulator.peek_observable_expectation(string) == 1
 
         # the state of the simulator is a +1 eigenstate of all logical Z operators
         for op in codes.QuditCode.get_logical_ops(code, Pauli.Z):
-            string = circuits.op_to_string(op)
+            string = op_to_string(op)
             assert simulator.peek_observable_expectation(string) == 1
 
         # the state of the simulator is a +1 eigenstate of all gauge Z operators
         for op in codes.QuditCode.get_gauge_ops(code, Pauli.Z):
-            string = circuits.op_to_string(op)
+            string = op_to_string(op)
             assert simulator.peek_observable_expectation(string) == 1
