+++ qldpc/math.py
@@ -24,8 +24,29 @@ import galois
 import numpy as np
 import numpy.typing as npt
 import scipy.special
+import stim
+
+from qldpc.objects import Pauli
 
 IntegerArray = TypeVar("IntegerArray", npt.NDArray[np.int_], galois.FieldArray)
+
+
+def op_to_string(op: npt.NDArray[np.int_]) -> stim.PauliString:
+    """Convert an integer array that represents a Pauli string into a stim.PauliString.
+
+    The (first, second) half the array indicates the support of (X, Z) Paulis.
+    """
+    support_xz = np.array(op, dtype=int).reshape(2, -1)
+    paulis = [Pauli((support_xz[0, qq], support_xz[1, qq])) for qq in range(support_xz.shape[1])]
+    return stim.PauliString(map(str, paulis))
+
+    num_qubits = len(op) // 2
+    paulis = ""
+    for qubit in range(num_qubits):
+        val_x = int(op[qubit])
+        val_z = int(op[qubit + num_qubits])
+        paulis = str(Pauli((val_x, val_z)))
+    return stim.PauliString(paulis)
 
 
 def symplectic_conjugate(vectors: IntegerArray) -> IntegerArray:
